chore(ci): Automatically prepare a PR for the release
# Issue

After creating a plugin release a manual PR on
`cloudfoundry/cli-plugin-repo` needed to be created to make it available
to consumers. This was cumbersome and error-prone.

# Fix

This PR adds follow-up steps to the release creation that create a
commit which can be used to create a PR on
`cloudfoundry/cli-plugin-repo`.

Unfortunately also creating the PR itself seems to be failing due to
https://github.com/cli/cli/issues/2691